I added custom hook useFilterUsersByText
